completely restructured the server code
